+++ src/cdogsed/nk_window.h
@@ -90,6 +90,8 @@ void DrawPic(
 void DrawTextbox(
 	struct nk_context *ctx, char *value, const int len, const char *tooltip,
 	const nk_flags flags);
-bool DrawNumberSlider(struct nk_context *ctx, const char *label, const char *tooltip, const int min, const int max, const float step, int *value);
+bool DrawNumberSlider(
+	struct nk_context *ctx, const char *label, const char *tooltip,
+	const int min, const int max, const int step, int *value);
 bool ColorPicker(
 	struct nk_context *ctx, const float height, const char *label, color_t *c);
